+++ nidx/nidx_vector/src/data_point_provider/mod.rs
@@ -19,27 +19,14 @@
 //
 
 pub mod reader;
-pub mod writer;
-
-pub mod state;
 
 use crate::config::VectorConfig;
 pub use crate::data_types::dtrie_ram::DTrie;
 use crate::data_types::DeleteLog;
 use crate::formula::Formula;
-use crate::VectorR;
-use std::fs::File;
-use std::io::{BufReader, BufWriter, Write};
-use std::path::Path;
 use std::time::SystemTime;
 
 pub use crate::data_point::Neighbour;
-
-const METADATA: &str = "metadata.json";
-const STATE: &str = "state.bincode";
-const TEMP_STATE: &str = "temp_state.bincode";
-const OPENING_FLAG: &str = ".opening";
-const WRITING_FLAG: &str = ".writing";
 
 pub trait SearchRequest {
     fn get_query(&self) -> &[f32];
@@ -57,22 +44,5 @@ struct TimeSensitiveDLog<'a> {
 impl<'a> DeleteLog for TimeSensitiveDLog<'a> {
     fn is_deleted(&self, key: &[u8]) -> bool {
         self.dlog.get(key).map(|t| t > self.time).unwrap_or_default()
-    }
-}
-
-struct IndexMetadata;
-impl IndexMetadata {
-    pub fn write(config: &VectorConfig, path: &Path) -> VectorR<()> {
-        let mut writer = BufWriter::new(File::create(path.join(METADATA))?);
-        serde_json::to_writer(&mut writer, config)?;
-        Ok(writer.flush()?)
-    }
-    pub fn open(path: &Path) -> VectorR<Option<VectorConfig>> {
-        let path = &path.join(METADATA);
-        if !path.is_file() {
-            return Ok(None);
-        }
-        let mut reader = BufReader::new(File::open(path)?);
-        Ok(Some(serde_json::from_reader(&mut reader)?))
     }
 }
