+++ nidx/nidx_vector/src/lib.rs
@@ -19,41 +19,34 @@
 //
 
 use config::VectorConfig;
-use data_point::{open, DataPointPin};
+use data_point::open;
 use data_point_provider::reader::TimeSensitiveDLog;
+use indexer::index_resource;
 use nidx_protos::Resource;
-use nidx_types::Seq;
-use service::VectorWriterService;
-use std::path::{Path, PathBuf};
+use nidx_types::{SegmentMetadata, Seq};
+use std::path::Path;
 use std::time::{Duration, SystemTime};
 
 pub struct VectorIndexer;
 
 impl VectorIndexer {
     pub fn index_resource(&self, output_dir: &Path, resource: &Resource) -> VectorR<(i64, Vec<String>)> {
-        let tmp = tempfile::tempdir()?;
-
         // Index resource
-        let mut writer = VectorWriterService::create(&tmp.path().join("index"), VectorConfig::default()).unwrap();
-        writer.set_resource(resource.into()).unwrap();
+        let (segment, deletions) = index_resource(resource.into(), output_dir, &VectorConfig::default()).unwrap();
 
-        // Copy just the segment to the output directory
-        let segments = writer.get_segment_ids().unwrap();
-        if segments.is_empty() {
-            return Ok((0, resource.sentences_to_delete.clone()));
+        if let Some(segment) = segment {
+            Ok((segment.records as i64, deletions))
+        } else {
+            Ok((0, deletions))
         }
-        assert!(segments.len() <= 1, "Expected a single segment");
-        std::fs::rename(tmp.path().join("index").join(&segments[0]), output_dir)?;
-
-        Ok((writer.count().unwrap() as i64, resource.sentences_to_delete.clone()))
     }
 
     pub fn merge(
         &self,
         work_dir: &Path,
-        segments: &[(PathBuf, Seq, i64)],
+        segments: Vec<(SegmentMetadata, Seq)>,
         deletions: &Vec<(Seq, &Vec<String>)>,
-    ) -> VectorR<(String, usize)> {
+    ) -> VectorR<usize> {
         // TODO: Maybe segments should not get a DTrie of deletions and just a hashset of them, and we can handle building that here?
         // Wait and see how the Tantivy indexes turn out
         let mut delete_log = data_point_provider::DTrie::new();
@@ -66,19 +59,13 @@ impl VectorIndexer {
 
         // Rename (nidx_vector wants uuid, we use i64 as segment ids) and open the segments
         let segment_ids: Vec<_> = segments
-            .iter()
-            .map(|(segment_path, seq, _)| {
-                let uuid = uuid::Uuid::new_v4();
-                std::fs::rename(segment_path, work_dir.join(uuid.to_string())).unwrap();
-                (uuid, seq)
-            })
-            .map(|(dpid, seq)| {
-                let dp = data_point::DataPointPin::open_pin(work_dir, dpid).unwrap();
-                let open_dp = open(&dp).unwrap();
+            .into_iter()
+            .map(|(meta, seq)| {
+                let open_dp = open(meta).unwrap();
                 (
                     TimeSensitiveDLog {
                         dlog: &delete_log,
-                        time: SystemTime::UNIX_EPOCH + Duration::from_secs(i64::from(seq) as u64),
+                        time: SystemTime::UNIX_EPOCH + Duration::from_secs(i64::from(&seq) as u64),
                     },
                     open_dp,
                 )
@@ -86,15 +73,13 @@ impl VectorIndexer {
             .collect();
 
         // Do the merge
-        let destination = DataPointPin::create_pin(work_dir).unwrap();
         let open_destination = data_point::merge(
-            &destination,
+            work_dir,
             &segment_ids.iter().map(|(a, b)| (a, b)).collect::<Vec<_>>(),
             &VectorConfig::default(),
-            SystemTime::now(),
         )?;
 
-        Ok((destination.id().to_string(), open_destination.journal().no_nodes()))
+        Ok(open_destination.no_nodes())
     }
 }
 
@@ -106,18 +91,15 @@ pub mod data_point;
 pub mod data_point_provider;
 mod data_types;
 pub mod formula;
+pub mod indexer;
+mod query_io;
 pub mod query_language;
-pub mod service;
 mod utils;
 mod vector_types;
 
 use thiserror::Error;
 #[derive(Debug, Error)]
 pub enum VectorErr {
-    #[error("Error using bincode: {0}")]
-    BincodeError(#[from] bincode::Error),
-    #[error("json error: {0}")]
-    SJ(#[from] serde_json::Error),
     #[error("IO error: {0}")]
     IoErr(#[from] std::io::Error),
     #[error("This index does not have an alive writer")]
