+++ nidx/nidx_vector/src/query_io.rs
